# Performance live heap profiling - deallocation code path