--- a/include/ddprof_stats.hpp
+++ b/include/ddprof_stats.hpp
@@ -14,6 +14,7 @@
   X(EVENT_COUNT, "event.count", STAT_GAUGE)                                    \
   X(EVENT_LOST, "event.lost", STAT_GAUGE)                                      \
   X(SAMPLE_COUNT, "sample.count", STAT_GAUGE)                                  \
+  X(UNMATCHED_DEALLOCATION_COUNT, "unmatched_deallocation.count", STAT_GAUGE)  \
   X(TARGET_CPU_USAGE, "target_process.cpu_usage.millicores", STAT_GAUGE)       \
   X(UNWIND_AVG_TIME, "unwind.avg_time_ns", STAT_GAUGE)                         \
   X(UNWIND_FRAMES, "unwind.frames", STAT_GAUGE)                                \