--- a/include/lib/address_bitset.hpp
+++ b/include/lib/address_bitset.hpp
@@ -7,13 +7,20 @@
 
 namespace ddprof {
 class AddressBitset {
+  // Number of bits is the number of addresses we can store
+  // We have one address per individual bit).
+  // so lets say you have 1111, you can store 4 addresses.
+  // We hash the address to a number (to have an equal probability of using
+  // all bits). Then we use the mask to position this address in our bitset.
+  // Addr -> Hash -> Mask (to get useable bits) -> Position in the bitset
+  // Note: the hashing step might be bad for cache locality.
 public:
   // returns true if the element was inserted
   bool set(uintptr_t addr);
   // returns true if the element was removed
   bool unset(uintptr_t addr);
   void clear() {
-    for (auto& element : _address_bitset) {
+    for (auto &element : _address_bitset) {
       element.store(0, std::memory_order_relaxed);
     }
     _nb_elements.store(0);
@@ -22,6 +29,8 @@ public:
 
 private:
   // 1 Meg divided in uint64's size
+  // The probability of collision is proportional to the number of elements
+  // already within the bitset
   constexpr static unsigned _nb_bits = 8 * 1024 * 1024;
   constexpr static unsigned _k_nb_elements = (_nb_bits) / (64);
   constexpr static unsigned _nb_bits_mask = _nb_bits - 1;