--- a/include/lib/address_bitset.hpp
+++ b/include/lib/address_bitset.hpp
@@ -1,3 +1,7 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0. This product includes software
+// developed at Datadog (https://www.datadoghq.com/). Copyright 2021-Present
+// Datadog, Inc.
 #pragma once
 
 #include <array>