--- a/src/jit/jitdump.cc
+++ b/src/jit/jitdump.cc
@@ -7,6 +7,8 @@
 #include <span>
 #include <vector>
 
+#define DEBUG
+
 // If we want to consider big endian, we will need this
 // has bswap_64/32/16
 // #include <byteswap.h>