# What does this PR do?

Ensure the structure to keep the allocation addresses is lock free.
The proposal is to have a bitset (using atomic integers) instead of a hash map. There is some logic to map the address to a position in the bitset.

# Motivation

Performances:
I get significantly better numbers in the benchmark (when focusing on the allocation code paths)

```
# New numbers are in the same ball park
BM_ShortLived_NoTracking/process_time/real_time     440168 ns       998071 ns         1349
BM_ShortLived_Tracking/process_time/real_time       490663 ns      1160473 ns         1505
BM_LongLived_NoTracking/process_time                345291 ns       218756 ns         3223
BM_LongLived_Tracking/process_time                  342495 ns       358709 ns         1840

# Previous numbers (with same config)
BM_ShortLived_NoTracking/process_time/real_time     448681 ns       946333 ns         1602
BM_ShortLived_Tracking/process_time/real_time      1860407 ns      5558747 ns          377
BM_LongLived_NoTracking/process_time                339814 ns       197875 ns         3370
BM_LongLived_Tracking/process_time                 1316344 ns      4319851 ns          142
```

# Additional Notes

The bitset has a limited number of elements. This impacts the probability of capturing the allocation event.
This only occurs when live allocation is triggered.

# How to test the change?

The improvement in performances can be observed through benchmarks.
The unit test cover the new data structure.
The existing tests already test live heap profiling.

Thanks to @richardstartin for mentioning this data structure.